requirements.txt
aiohttp==3.8.4
asyncio==3.4.3
python-environ==0.4.54
asyncpg==0.27.0
SQLAlchemy==2.0.9
greenlet==2.0.2